pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rocmate"
version = "0.1.0"
description = "Curated AMD GPU compatibility index and CLI for AI workloads"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Benjamin Faeuster", email = "benjamin.faeuster@web.de" }]
keywords = ["rocm", "amd", "gpu", "ai", "ollama", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: GPU",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Hardware",
]
dependencies = [
    "typer>=0.12",
    "rich>=13.7",
    "pydantic>=2.6",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "mypy>=1.10",
]

[project.scripts]
rocmate = "rocmate.cli:app"

[project.urls]
Homepage = "https://github.com/tonde/rocmate"
Issues = "https://github.com/tonde/rocmate/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/rocmate"]
# Include YAML configs in the wheel
[tool.hatch.build.targets.wheel.force-include]
"configs" = "rocmate/_configs"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=rocmate --cov-report=term-missing"
